Remove ParticleToMeshMap, it's just an aspect of SpatialMesh
It is already very tightly coupled with SpatialMesh,
does not make sense to keep it a separate class.
