+++ SpatialMesh.py
@@ -2,6 +2,7 @@ import logging
 
 import numpy as np
 
+from Vec3d import Vec3d
 from ef.util.serializable_h5 import SerializableH5
 
 
@@ -66,6 +67,28 @@ class SpatialMesh(SerializableH5):
                             f"{self.cell[i]:.3f} from {step_size[i]:.3f} "
                             f"to fit in a round number of cells.")
         return self
+
+    def weight_particles_charge_to_mesh(self, particle_sources):
+        volume_around_node = self.cell.prod()
+        for part_src in particle_sources.sources:
+            for p in part_src.particles:
+                charge = p.charge / volume_around_node
+                node, remainder = np.divmod(p._position, self.cell)
+                nx, ny, nz = np.array(node, np.int)
+                cell_slice = self.charge_density[nx:nx + 2, ny:ny + 2, nz:nz + 2]
+                weight = remainder / self.cell
+                weights = np.array([1. - weight, weight])
+                dn = np.moveaxis(np.mgrid[0:2, 0:2, 0:2], 0, -1)
+                cell_slice += weights[dn, [0, 1, 2]].prod(-1) * charge
+
+    def field_at_position(self, position):
+        node, remainder = np.divmod(position, self.cell)
+        nx, ny, nz = np.array(node, np.int)
+        cell_slice = self.electric_field[nx:nx + 2, ny:ny + 2, nz:nz + 2]
+        weight = remainder / self.cell
+        weights = np.array([1. - weight, weight])
+        dn = np.moveaxis(np.mgrid[0:2, 0:2, 0:2], 0, -1)
+        return Vec3d(*(weights[dn, [0, 1, 2]].prod(-1)[:, :, :, np.newaxis] * cell_slice).sum((0, 1, 2)))
 
     def clear_old_density_values(self):
         self.charge_density.fill(0)
